# internal/api/openapi.yaml
basePath: /api
definitions:
...
  api.addMemberRequest:
    properties:
      email:
        type: string
      role:
        type: string
    type: object
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.memberResponse:
    properties:
      addedAt:
        type: string
      addedBy:
        type: string
      email:
        type: string
      role:
        type: string
    type: object
...
  api.updateMemberRequest:
    properties:
      role:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: Self-hosted Railway-style deploy platform for Kubernetes. Manages projects,
    apps, deployments, env vars, secrets, domains, and git providers.
  title: Mortise API
  version: 0.1.0
paths:
...
  /projects/{project}/members:
    get:
      description: Returns all members of a project with their roles
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.memberResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: List project members
      tags:
      - members
    post:
      consumes:
      - application/json
      description: Adds a user as a member of a project with the specified role
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: Member email and role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.addMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.memberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Add a project member
      tags:
      - members
  /projects/{project}/members/{email}:
    delete:
      description: Removes a member from a project. Cannot remove the last owner.
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: Member email
        in: path
        name: email
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Remove a project member
      tags:
      - members
    patch:
      consumes:
      - application/json
      description: Changes the role of an existing project member
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: Member email
        in: path
        name: email
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.memberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a project member's role
      tags:
      - members
...
securityDefinitions:
  BearerAuth:
    description: JWT bearer token from POST /api/auth/login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"